requirements.txt
astroid==2.5.1
autopep8==1.5.5
certifi==2020.12.5
chromedriver-binary-auto==0.1.1
colorama==0.4.4
isort==5.7.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
playsound==1.2.2
pycodestyle==2.6.0
pylint==2.7.2
python-vlc==3.0.12117
selenium==3.141.0
toml==0.10.2
urllib3==1.26.4
wincertstore==0.2
wrapt==1.12.1
git+https://github.com/pgoelter/pyconfig